chore: add codelyzer
